src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/distribution-update-recalcs-style-expected.txt

CONSOLE ERROR: line 27: Uncaught TypeError: document.getElementById(...).createShadowRoot is not a function
Tests to ensure that updating the select rule in projection causes style recalc.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".




    

        

        
    







FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
